Bake Vertex Lighting
We now bake the blender lighting to materials that have "VCol Light"
enabled. In that case, manual lighting will be ignored. If you want the
old functionality back, disable "VCol Light". The result in Plasma should
be very close to the "Shaded" view. If you wish to include dynamic
lighting (for the avatar), you will need to use light groups.